+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 * Fixed an issue where the ***generate-docs*** command reset the enumeration of line numbering after an MD table.
 * Updated the **upload** command to support mappers.
+* Fixed an issue where integration with v2 were recognised by the `name` field instead of the `display` field in the **validate** command.
 * Added a playbook validation to check if a task script exists in the id set in the **validate** command.
 
 # 1.2.10